Project_1_Forecasting_App-main/project_1_Forecasting_App/Final_App/requirements.txt
alpaca-trade-api==2.2.0
python-dotenv==0.20.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.2
requests==2.27.1
streamlit==1.10.0
